# Conflicts:
#	src/main/java/com/example/workhive/domain/dto/MemoDTO.java
#	src/main/java/com/example/workhive/domain/entity/ChatMessageEntity.java
#	src/main/resources/templates/member/idCheck.html
#	src/main/resources/templates/member/joinForm.html
#	src/main/resources/templates/member/loginForm.html

+++ src/main/java/com/example/workhive/controller/ChatMessageController.java
@@ -18,7 +18,7 @@ public class ChatMessageController {
     }
 
     @GetMapping("/{chatRoomId}")
-    public ResponseEntity<List<ChatMessageDTO>> getMessages(@PathVariable("id") Integer chatRoomId) {
+    public ResponseEntity<List<ChatMessageDTO>> getMessages(@PathVariable("id") Long chatRoomId) {
         List<ChatMessageDTO> messages = chatMessageService.getMessagesByChatRoom(chatRoomId);
         return ResponseEntity.ok(messages);
     }
